sparse embedding implementation

+++ svm.cpp
@@ -11,17 +11,17 @@ void LinearSVM(const std::vector<std::vector<double>*>& feature, const std::vect
     double pos_penalty, double neg_penalty, std::vector<double>* coeff) {
     int dim = feature[0]->size();
     std::vector<double> w(dim, 0);
-    for (int i = 0; i < coeff->size(); ++i)
+    for (int i = 0; i < (int)coeff->size(); ++i)
         if (coeff->at(i) > 0)
         for (int j = 0; j < dim; ++j)
             w[j] += label[i] * feature[i]->at(j) * coeff->at(i);
     std::vector<double> Q(coeff->size(), 0);
-    for (int i = 0; i < coeff->size(); ++i)
+    for (int i = 0; i < (int)coeff->size(); ++i)
         for (int j = 0; j < dim; ++j)
             Q[i] += sqr(feature[i]->at(j));
 
     std::vector<int> order(coeff->size());
-    for (int i = 0; i < coeff->size(); ++i)
+    for (int i = 0; i < (int)coeff->size(); ++i)
         order[i] = i;
     for (int epoch = 0; epoch < LINEAR_EPOCHS; ++epoch) {
         RandomPermutation(&order);
